Adopt UIScene lifecycle to fix launch crash on iOS 27
Apps built with the iOS 27 SDK must adopt the UIScene lifecycle;
OTPKitDemo crashed at launch with EXC_BREAKPOINT in
__UIApplicationEvaluateRuntimeIssueForNoSceneLifecycleAdoption because
it created its window in the app delegate with no scene delegate.

- Add SceneDelegate owning the window and root view controller
  selection (onboarding vs. main), using UIWindow(windowScene:)
- Slim AppDelegate to a bare UIApplicationDelegate
- Declare a static UIApplicationSceneManifest in OTPKitDemo-Info.plist
  and drop the INFOPLIST_KEY_UIApplicationSceneManifest_Generation
  build settings so the manifest has a single source of truth
